requirements.txt
Flask==1.0.2
google-api-core
google-auth
google-cloud-core
google-cloud-ndb
google-cloud-storage
mock